Umount a drive when it is Released
Currently, when we release a drive using `kubectl direct-csi drive release` command,
we would make the DriveStatus as 'Released'. And any `Released` drive wouldn't be
included in any of the direct-csi lifecycle.

This PR will umount a drive when it is in Released state and make it Available after the
umount succeeds