Support virtual disks by relaxing the hashed uuid generation for the drive names
currently, the name of the newly created drive is a uuid which will be calculated by
the hash of the hardware persistent properties of the drive. This prevents creation of
any virtual disks which will not have any ids on them.

by assigning random uuids for the drive names, we acccept such virtual disks

Fixes #580